+++ relay-log/src/setup.rs
@@ -7,6 +7,9 @@ use serde::{Deserialize, Serialize};
 use tracing::{level_filters::LevelFilter, Level};
 use tracing_subscriber::{prelude::*, EnvFilter, Layer};
 
+#[cfg(feature = "dashboard")]
+use crate::LOGS;
+
 /// The full release name including the Relay version and SHA.
 const RELEASE: &str = std::env!("RELAY_RELEASE");
 
@@ -196,6 +199,29 @@ fn get_default_filters() -> EnvFilter {
     env_filter
 }
 
+#[cfg(feature = "dashboard")]
+struct LogsWriter;
+
+#[cfg(feature = "dashboard")]
+impl std::io::Write for LogsWriter {
+    fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
+        let tx = &*LOGS;
+        let buf_len = buf.len();
+        tx.send(buf.to_vec()).ok();
+        Ok(buf_len)
+    }
+
+    fn flush(&mut self) -> std::io::Result<()> {
+        Ok(())
+    }
+}
+
+/// Returns the log writer.
+#[cfg(feature = "dashboard")]
+fn make_logs_writer() -> impl std::io::Write {
+    LogsWriter
+}
+
 /// Initialize the logging system and reporting to Sentry.
 ///
 /// # Example
@@ -219,6 +245,13 @@ pub fn init(config: &LogConfig, sentry: &SentryConfig) {
         .with_writer(std::io::stderr)
         .with_target(true);
 
+    #[cfg(feature = "dashboard")]
+    let dashboard_subscriber = tracing_subscriber::fmt::layer()
+        .with_writer(make_logs_writer)
+        .with_target(true)
+        .with_ansi(true)
+        .compact();
+
     let format = match (config.format, console::user_attended()) {
         (LogFormat::Auto, true) | (LogFormat::Pretty, _) => {
             subscriber.compact().without_time().boxed()
@@ -236,14 +269,19 @@ pub fn init(config: &LogConfig, sentry: &SentryConfig) {
             .boxed(),
     };
 
-    tracing_subscriber::registry()
+    let trsub = tracing_subscriber::registry()
         .with(format.with_filter(LevelFilter::from(config.level)))
         .with(sentry::integrations::tracing::layer())
         .with(match env::var(EnvFilter::DEFAULT_ENV) {
             Ok(value) => EnvFilter::new(value),
             Err(_) => get_default_filters(),
-        })
-        .init();
+        });
+
+    // Also add dashboard subscriber if the feature is enabled.
+    #[cfg(feature = "dashboard")]
+    let trsub = trsub.with(dashboard_subscriber);
+
+    trsub.init();
 
     if let Some(dsn) = sentry.enabled_dsn() {
         let guard = sentry::init(sentry::ClientOptions {